# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super Application API",
        "version": "8.1.6"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/user/index/{role}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users by role",
                "description": "Retrieves a list of users based on the specified role.",
                "operationId": "3b45645a18603a9863a59011d7dbc746",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The role of the users to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "user"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "role_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "12345678"
                                                    },
                                                    "directorate_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "city_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "flag": {
                                                        "type": "string",
                                                        "example": "0"
                                                    },
                                                    "certificate": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "example": [
                                                            "certificate1",
                                                            "certificate2"
                                                        ],
                                                        "nullable": true
                                                    },
                                                    "gender": {
                                                        "type": "string",
                                                        "example": "male"
                                                    },
                                                    "courses": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "example": [
                                                            "course1",
                                                            "course2"
                                                        ],
                                                        "nullable": true
                                                    },
                                                    "city_name": {
                                                        "type": "string",
                                                        "example": "city one"
                                                    },
                                                    "directorate_name": {
                                                        "type": "string",
                                                        "example": "directorate one"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/show/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user details",
                "description": "Retrieves details of a specific user.",
                "operationId": "964147b22536da0774f2ca4e1a8fb569",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "role_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "admin"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "admin"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "12345678"
                                                },
                                                "flag": {
                                                    "type": "string",
                                                    "example": "0"
                                                },
                                                "certificate": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "nullable": true
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "male"
                                                },
                                                "courses": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "nullable": true
                                                },
                                                "city_name": {
                                                    "type": "string",
                                                    "example": "city one"
                                                },
                                                "directorate_name": {
                                                    "type": "string",
                                                    "example": "directorate one"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/user/stopaccount/{user}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Stop user account",
                "description": "Stops the account of a specific user.",
                "operationId": "a855ca9a04c27c64893a58f5591e726e",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user account stopped successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}